+++ .gitignore
@@ -1,0 +1,4 @@
+# Project exclude paths
+/out/
+*.xml
+*.iml
