+++ src/ru/skypro/Author.java
@@ -1,0 +1,17 @@
+package ru.skypro;
+
+public class Author {
+    final String name;
+    final String surName;
+    public Author(String name, String surName){
+        this.name = name;
+        this.surName = surName;
+    }
+
+    public String getName(){
+        return name;
+    }
+    public String getSurName(){
+        return surName;
+    }
+}
